moved redux
