+++ client/src/index.js
@@ -3,7 +3,7 @@ import { render } from 'react-dom';
 import { BrowserRouter } from 'react-router-dom';
 
 import { Provider } from 'react-redux';
-import store from './store';
+import store from './redux/store';
 
 import './index.css';
 import App from './App';
