+++ client/src/redux/reducers/views.js
